+++ Sources/WalletConnect_Swift/Serialiser/JSONRPCSerialiserError.swift
@@ -1,0 +1,10 @@
+// 
+
+import Foundation
+
+enum JSONRPCSerialiserError: String, Error, CustomStringConvertible {
+    case messageToShort = "Error: message is to short"
+    var description: String {
+        return rawValue
+    }
+}
