+++ Sources/WalletConnect_Swift/Types/ClientSynchronisation/PairingApproveParams.swift
@@ -1,0 +1,11 @@
+// 
+
+import Foundation
+
+struct PairingApproveParams: Codable, Equatable {
+    let topic: String
+    
+    enum CodingKeys: CodingKey {
+        case topic
+    }
+}
