+++ Sources/WalletConnect_Swift/Types/ClientSynchronisation/PairingRejectParams.swift
@@ -1,0 +1,11 @@
+// 
+
+import Foundation
+
+struct PairingRejectParams: Codable, Equatable {
+    let reason: String
+    
+    enum CodingKeys: CodingKey {
+        case reason
+    }
+}
